add arduino file, merge with cardriver

+++ src/com/example/RC_Car/IControlsHandler.java
@@ -4,6 +4,6 @@ package com.example.RC_Car;
  * Created by mikolaj on 05.01.15.
  */
 public interface IControlsHandler {
-    public double getSpeed();
-    public double getDirection();
+    public int getSpeed();
+    public int getDirection();
 }
